requirements.txt
Flask==2.3.3
python-docx==0.8.11
pdfplumber==0.10.3
PyPDF2==3.0.1
wordcloud==1.9.2
nltk==3.8.1
unidecode==1.3.7
matplotlib==3.8.2
Werkzeug==2.3.7
